+++ server/src/embedded_migrations.rs
@@ -133,4 +133,5 @@ pub const EMBEDDED_MIGRATIONS: &[(&str, &str)] = &[
     ("130_archive_duplicate_shipping_alteration_fee_products.sql", include_str!("../../migrations/130_archive_duplicate_shipping_alteration_fee_products.sql")),
     ("131_repair_counterpoint_historical_paid_amounts.sql", include_str!("../../migrations/131_repair_counterpoint_historical_paid_amounts.sql")),
     ("132_reliability_read_path_indexes.sql", include_str!("../../migrations/132_reliability_read_path_indexes.sql")),
+    ("133_qbo_backdated_sale_clearing.sql", include_str!("../../migrations/133_qbo_backdated_sale_clearing.sql")),
 ];